controllo sicurezza remove disp

+++ get2/calendario/ajax.py
@@ -189,8 +189,9 @@ def disp(request, turno_id, mansione_id, persona_id, disp):
 		dajax.assign('#disponibilita-'+str(p.id), 'innerHTML', temp.render(c))	
 	elif request.user.is_staff:
 		d=Disponibilita.objects.get(persona=p,turno=t)
-		d.delete()
-		dajax.assign('#disponibilita-'+str(p.id), 'innerHTML', '')
+		if request.user.is_superuser or request.user==d.creata_da:
+			d.delete()
+			dajax.assign('#disponibilita-'+str(p.id), 'innerHTML', '')
 	html_anteprima = render_to_string( 'turno.html', { 't': t, 'request':request } )
 	dajax.assign('div #anteprima', 'innerHTML', html_anteprima+'<div style="clear:both;"></div>')
 	dajax.script('$(".bottom-right").notify({ message: { text: "Aggiornata disponibilita per '+str(p)+'" }}).show();')
